+++ tests/wrappers/test_log1p_mel_spectrogram.py
@@ -4,9 +4,12 @@ import matplotlib.pyplot as plt
 import numpy as np
 from gym import spaces
 
-from pynktrombonegym.env import PynkTrombone
+from pynktrombonegym.environment import PynkTrombone
 from pynktrombonegym.spaces import ObservationSpaceNames as OSN
-from pynktrombonegym.spectrogram import calc_rfft_channel_num, calc_target_sound_spectrogram_length
+from pynktrombonegym.spectrogram import (
+    calc_rfft_channel_num,
+    calc_target_sound_spectrogram_length,
+)
 from pynktrombonegym.wrappers import log1p_mel_spectrogram as l1pms
 
 from ..test_env import assert_space, target_sound_files
